auth-app-home/requirements.txt
alembic==1.8.1
asyncpg==0.26.0
autopep8==1.7.0
dependency-injector==4.40.0
email-validator==1.3.0
fastapi==0.84.0
fastapi-sessions==0.3.2
httpx==0.23.0
itsdangerous==2.1.2
jinja2==3.1.2
passlib[bcrypt]==1.7.4
pylint==2.15
pytest==7.1.3
pytest-asyncio==0.19.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.5
redis==4.3.4
sqlalchemy[asyncio]==1.4.41
uvicorn[standard]==0.18.3
